+++ routes/games.js
@@ -1,13 +1,33 @@
 const gamesRouter = require("express").Router();
+
 const {
-  sendAllGames,
+  getAllGames,
+  checkIsTitleInArray,
+  updateGamesArray,
+  updateGamesFile,
+  findGameById,
   deleteGame,
-  addGameController,
-} = require("../controllers/games");
-const getAllGames = require("../middlewares/games");
+} = require("../middlewares/games");
+const { sendAllGames, sendUpdatedGames } = require("../controllers/games");
 
 gamesRouter.get("/games", getAllGames, sendAllGames);
-gamesRouter.delete("/games/:id", getAllGames, deleteGame);
-gamesRouter.post("/games", getAllGames, addGameController);
+
+gamesRouter.post(
+  "/games",
+  getAllGames,
+  checkIsTitleInArray,
+  updateGamesArray,
+  updateGamesFile,
+  sendUpdatedGames
+);
+
+gamesRouter.delete(
+  "/games/:id",
+  getAllGames,
+  findGameById,
+  deleteGame,
+  updateGamesFile,
+  sendUpdatedGames
+);
 
 module.exports = gamesRouter;
